Break further circular dependencies
