+++ ts/libs/core/src/lib/model/custom-resource-definition.prm.ts
@@ -1,4 +1,3 @@
-import { JSONSchema7 } from 'json-schema';
 import { OpenAPIV3 } from 'openapi-types';
 import { PolarisType } from '../transformation';
 import { initSelf } from '../util';
@@ -7,12 +6,6 @@ import { ObjectKind } from './object-kind.prm';
 
 /** Describes a type using an OpenAPI v3 schema. */
 export type OpenApiSchema = OpenAPIV3.BaseSchemaObject;
-
-/** Describes a type using JSON Schema. */
-export type JsonSchema<T = any> = Omit<JSONSchema7, 'properties' | 'items'> & {
-    properties?: { [K in keyof T]?: JsonSchema<T[K]>; };
-    items?: JsonSchema<T> | JsonSchema<T>[];
-};
 
 /**
  * Describes a custom {@link ApiObject} type.
