+++ ts/libs/core/src/lib/slo/public/common/index.ts
@@ -1,3 +1,2 @@
-export * from './errors';
 export * from './service-level-objective';
 export * from './slo-output';
