+++ ts/libs/core/src/lib/slo/public/control/index.ts
@@ -1,2 +1,3 @@
+export * from './errors';
 export * from './slo-control-loop';
 export * from './slo-evaluator';
