+++ ts/libs/core/src/lib/transformation/public/common/polaris-transformation-service.ts
@@ -1,5 +1,5 @@
-import { JsonSchema, ObjectKind } from '../../../model';
-import { PolarisConstructor } from '../../../util';
+import { ObjectKind } from '../../../model/object-kind.prm';
+import { JsonSchema, PolarisConstructor } from '../../../util';
 
 /**
  * Implementations of this service should be used for converting between orchestrator-independent Polaris objects
