+++ ts/libs/core/src/lib/transformation/public/impl/default-polaris-transformation.service.ts
@@ -1,5 +1,5 @@
-import { JsonSchema, ObjectKind } from '../../../model';
-import { IndexByKey, PolarisConstructor, cloneDeepWithoutExcluded } from '../../../util';
+import { ObjectKind } from '../../../model';
+import { IndexByKey, JsonSchema,  PolarisConstructor, cloneDeepWithoutExcluded } from '../../../util';
 import { PolarisMetadataUtils, PolarisTransformationMetadata } from '../../internal';
 import { PolarisTransformationConfig, PolarisTransformationServiceManager, PolarisTransformer, UnknownObjectKindError } from '../common';
 import { DefaultTransformer } from './transformers';
