+++ ts/libs/core/src/lib/transformation/public/impl/transformers/default-transformer.ts
@@ -1,7 +1,6 @@
 import { classToPlain, plainToClass } from 'class-transformer';
 import { cloneDeep } from 'lodash';
-import { JsonSchema } from '../../../../model';
-import { Constructor, InterfaceOf, cloneDeepWithoutExcluded } from '../../../../util';
+import { Constructor, InterfaceOf, JsonSchema, cloneDeepWithoutExcluded } from '../../../../util';
 import { PolarisTransformationService, PolarisTransformer } from '../../common';
 
 /**
